change columns with string values to `object` and not categoricals
seems `object` is the better default for strings, even though categorical highlighted by being first and in lots of examples in current documentation. Found mention of `object` in a highlight box down below eventually. Having as categorical causes problems with groupby later if subset dataframes. Solution to when need to use categoricals is to add `observed=True` to groupby, but unncessary of don't absolutely need column values to be categorical in first place.